FullStack/backend/ai_service/requirements.txt
fastapi>=0.110.1
uvicorn>=0.24.0
chromadb>=0.4.0
google-generativeai>=0.8.6,<0.9.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
